backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
alpaca-py==0.33.0
yfinance==0.2.50
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
websockets==13.1
pydantic==2.10.3
httpx==0.28.1
